+++ lib/ble_peripheral.dart
@@ -1,5 +1,11 @@
 library ble_peripheral;
 
 export 'package:ble_peripheral/src/ble_peripheral.dart';
-export 'package:ble_peripheral/src/models/ble_enums.dart';
-export 'package:ble_peripheral/src/generated/ble_peripheral.g.dart';
+export 'package:ble_peripheral/src/generated/ble_peripheral.g.dart'
+    show
+        BleService,
+        BleCharacteristic,
+        CharacteristicProperties,
+        AttributePermissions,
+        SubscribedClient,
+        BondState;