+++ windows/ble_peripheral_plugin.cpp
@@ -19,7 +19,6 @@ namespace ble_peripheral
   using ble_peripheral::ErrorOr;
   std::unique_ptr<BleCallback> bleCallback;
   std::map<std::string, GattServiceProviderObject *> serviceProviderMap;
-  std::mutex cout_mutex;
 
   // static
   void BlePeripheralPlugin::RegisterWithRegistrar(flutter::PluginRegistrarWindows *registrar)
@@ -36,23 +35,37 @@ namespace ble_peripheral
 
   winrt::fire_and_forget BlePeripheralPlugin::InitializeAdapter()
   {
-    auto radios = co_await Radio::GetRadiosAsync();
-    for (auto &&radio : radios)
+    const auto &bluetooth_adapter = co_await BluetoothAdapter::GetDefaultAsync();
+    if (bluetooth_adapter != nullptr)
     {
-      if (radio.Kind() == RadioKind::Bluetooth)
+      adapter = bluetooth_adapter;
+      bluetoothRadio = co_await bluetooth_adapter.GetRadioAsync();
+    }
+    else
+    {
+      std::cout << "Bluetooth adapter is not available" << std::endl;
+      auto radios = co_await Radio::GetRadiosAsync();
+      for (auto &&radio : radios)
       {
-        bluetoothRadio = radio;
-        radioStateChangedRevoker = bluetoothRadio.StateChanged(winrt::auto_revoke, {this, &BlePeripheralPlugin::Radio_StateChanged});
-        bool isOn = bluetoothRadio.State() == RadioState::On;
-        uiThreadHandler_.Post([isOn]
-                              { bleCallback->OnBleStateChange(isOn, SuccessCallback, ErrorCallback); });
-
-        break;
+        if (radio.Kind() == RadioKind::Bluetooth)
+        {
+          std::cout << "Bluetooth Radio found" << std::endl;
+          bluetoothRadio = radio;
+          break;
+        }
       }
     }
-    if (!bluetoothRadio)
+
+    if (bluetoothRadio != nullptr)
     {
-      std::cout << "Bluetooth is not available" << std::endl;
+      radioStateChangedRevoker = bluetoothRadio.StateChanged(winrt::auto_revoke, {this, &BlePeripheralPlugin::Radio_StateChanged});
+      bool isOn = bluetoothRadio.State() == RadioState::On;
+      uiThreadHandler_.Post([isOn]
+                            { bleCallback->OnBleStateChange(isOn, SuccessCallback, ErrorCallback); });
+    }
+    else
+    {
+      std::cout << "Bluetooth radio is not available" << std::endl;
     }
   }
 
@@ -64,17 +77,21 @@ namespace ble_peripheral
 
   ErrorOr<std::optional<bool>> BlePeripheralPlugin::IsAdvertising()
   {
-    // Check is any service is advertising, or if services list is empty
-    // Get serviceProviderMap length
-    if (serviceProviderMap.size() == 0)
+    if (!publisher)
+    {
       return ErrorOr<std::optional<bool>>(std::optional<bool>(false));
-    bool advertising = AreAllServicesStarted();
+    }
+    bool advertising = publisher.Status() == BluetoothLEAdvertisementPublisherStatus::Started;
     return ErrorOr<std::optional<bool>>(std::optional<bool>(advertising));
   };
 
   ErrorOr<bool> BlePeripheralPlugin::IsSupported()
   {
-    return bluetoothRadio != nullptr;
+    if (!adapter)
+    {
+      return FlutterError("Bluetooth adapter is not available");
+    }
+    return adapter.IsPeripheralRoleSupported();
   };
 
   ErrorOr<bool> BlePeripheralPlugin::AskBlePermission()
@@ -85,6 +102,14 @@ namespace ble_peripheral
 
   std::optional<FlutterError> BlePeripheralPlugin::AddService(const BleService &service)
   {
+    // Check if service already exists
+    std::string serviceId = to_lower_case(service.uuid());
+    if (serviceProviderMap.find(serviceId) != serviceProviderMap.end())
+    {
+      return FlutterError("Service already added");
+    }
+
+    // Add service
     AddServiceAsync(service);
     return std::nullopt;
   };
@@ -125,6 +150,43 @@ namespace ble_peripheral
     return services;
   }
 
+  ErrorOr<flutter::EncodableList> BlePeripheralPlugin::GetSubscribedClients()
+  {
+    // Map of deviceId to list of services and characteristics
+    auto deviceMap = std::map<std::string, flutter::EncodableList>();
+    for (auto const &[key, gattServiceObject] : serviceProviderMap)
+    {
+      for (auto const &[charKey, gattChar] : gattServiceObject->characteristics)
+      {
+        auto clients = gattChar->stored_clients;
+        for (unsigned int i = 0; i < clients.Size(); ++i)
+        {
+          auto client = clients.GetAt(i);
+          std::string deviceId = ParseBluetoothClientId(client.Session().DeviceId().Id());
+
+          if (deviceMap.find(deviceId) != deviceMap.end())
+          {
+            deviceMap[deviceId].push_back(flutter::EncodableValue(charKey));
+          }
+          else
+          {
+            flutter::EncodableList charList = flutter::EncodableList();
+            charList.push_back(flutter::EncodableValue(charKey));
+            deviceMap.insert_or_assign(deviceId, charList);
+          }
+        }
+      }
+    }
+
+    // Get all subscribed clients
+    flutter::EncodableList clients_list = flutter::EncodableList();
+    for (auto const &[deviceId, charList] : deviceMap)
+    {
+      clients_list.push_back(flutter::CustomEncodableValue(SubscribedClient(deviceId, charList)));
+    }
+    return clients_list;
+  }
+
   std::optional<FlutterError> BlePeripheralPlugin::StartAdvertising(
       const flutter::EncodableList &services,
       const std::string *local_name,
@@ -134,63 +196,75 @@ namespace ble_peripheral
   {
     try
     {
-      // check if services are empty
-      if (serviceProviderMap.size() == 0)
-        return FlutterError("No services added to advertise");
-
-      if (AreAllServicesStarted())
+      // Lazy initialization of publisher
+      if (!publisher)
       {
-        std::cout << "All services already advertising" << std::endl;
-        uiThreadHandler_.Post([]
-                              { bleCallback->OnAdvertisingStatusUpdate(true, nullptr, SuccessCallback, ErrorCallback); });
-        return std::nullopt;
+        publisher = BluetoothLEAdvertisementPublisher();
+        publisher.StatusChanged({this, &BlePeripheralPlugin::Publisher_StatusChanged});
       }
 
-      auto advertisementParameter = GattServiceProviderAdvertisingParameters();
-      advertisementParameter.IsDiscoverable(true);
-      advertisementParameter.IsConnectable(true);
-
-      for (auto const &[key, gattServiceObject] : serviceProviderMap)
+      if (publisher.Status() == BluetoothLEAdvertisementPublisherStatus::Started)
       {
-        if (gattServiceObject->obj.AdvertisementStatus() == GattServiceProviderAdvertisementStatus::Started)
-        {
-          std::cout << "Service " << key << " is already advertising, skipping" << std::endl;
-          continue;
-        }
-        gattServiceObject->obj.StartAdvertising(advertisementParameter);
+        return FlutterError("Already advertising");
       }
 
+
+      if (manufacturer_data != nullptr)
+      {
+        const auto leManufacturerData = BluetoothLEManufacturerData();
+        auto manufacturerId = static_cast<uint16_t>(manufacturer_data->manufacturer_id());
+        auto manufacturerBytes = from_bytevc(manufacturer_data->data());
+        leManufacturerData.CompanyId(manufacturerId);
+        leManufacturerData.Data(manufacturerBytes);
+        publisher.Advertisement().ManufacturerData().Append(leManufacturerData);
+      }
+
+      publisher.Start();
       return std::nullopt;
     }
+    catch (const winrt::hresult_error &e)
+    {
+      std::wcerr << "Failed with error: Code: " << e.code() << "Message: " << e.message().c_str() << std::endl;
+      std::string errorMessage = winrt::to_string(e.message());
+      return FlutterError(winrt::to_string(e.message()));
+    }
+    catch (const std::exception &e)
+    {
+      std::cout << "Error: " << e.what() << std::endl;
+      std::wstring errorMessage = winrt::to_hstring(e.what()).c_str();
+      return FlutterError(winrt::to_string(errorMessage));
+    }
+
     catch (...)
     {
-      std::cout << "Error: "
-                << "Unknown error" << std::endl;
-      return std::nullopt;
+      std::cout << "Error: Unknown error" << std::endl;
+      return FlutterError("Something Went Wrong");
     }
   }
 
   std::optional<FlutterError> BlePeripheralPlugin::StopAdvertising()
   {
-    for (auto const &[key, gattServiceObject] : serviceProviderMap)
+    try
     {
-      try
+      if (publisher)
       {
-        gattServiceObject->obj.StopAdvertising();
-        std::cout << "Stopped advertising for service: " << key << std::endl;
+        std::cout << "Stopping advertising" << std::endl;
+        publisher.Advertisement().ServiceUuids().Clear();
+        publisher.Advertisement().ManufacturerData().Clear();
+        publisher.Stop();
       }
-      catch (const winrt::hresult_error &e)
-      {
-        std::wcerr << "Failed to stop service: " << key.c_str() << ", Error: " << e.message().c_str() << std::endl;
-      }
-      catch (...)
-      {
-        std::cout << "Error: Unknown error" << std::endl;
-      }
+      return std::nullopt;
     }
-    uiThreadHandler_.Post([]
-                          { bleCallback->OnAdvertisingStatusUpdate(false, nullptr, SuccessCallback, ErrorCallback); });
-    return std::nullopt;
+    catch (const winrt::hresult_error &e)
+    {
+      std::wcerr << "Failed to stop advertise Error: " << e.message().c_str() << std::endl;
+      return FlutterError("Failed", winrt::to_string(e.message()));
+    }
+    catch (...)
+    {
+      std::cout << "Error: Unknown error" << std::endl;
+      return FlutterError("Failed", "Something Went Wrong");
+    }
   }
 
   std::optional<FlutterError> BlePeripheralPlugin::UpdateCharacteristic(
@@ -206,16 +280,52 @@ namespace ble_peripheral
     DataWriter writer;
     writer.ByteOrder(ByteOrder::LittleEndian);
     writer.WriteBuffer(bytes);
+
+    if (device_id != nullptr)
+    {
+      std::string deviceId = *device_id;
+      for (auto const &client : gattCharacteristicObject->stored_clients)
+      {
+        if (ParseBluetoothClientId(client.Session().DeviceId().Id()) == deviceId)
+        {
+          gattCharacteristicObject->obj.NotifyValueAsync(writer.DetachBuffer(), client);
+          return std::nullopt;
+        }
+      }
+      return FlutterError("Device not subscribed to this characteristic");
+    }
+
     gattCharacteristicObject->obj.NotifyValueAsync(writer.DetachBuffer());
     return std::nullopt;
   }
 
   // Helpers
+  //
+  void BlePeripheralPlugin::Publisher_StatusChanged(BluetoothLEAdvertisementPublisher const &sender, IInspectable const &args)
+  {
+    auto status = sender.Status();
+
+    if (status == BluetoothLEAdvertisementPublisherStatus::Started)
+    {
+      uiThreadHandler_.Post([]
+                            { bleCallback->OnAdvertisingStatusUpdate(true, nullptr, SuccessCallback, ErrorCallback); });
+    }
+    else if (status == BluetoothLEAdvertisementPublisherStatus::Stopped || status == BluetoothLEAdvertisementPublisherStatus::Aborted)
+    {
+      uiThreadHandler_.Post([]
+                            { bleCallback->OnAdvertisingStatusUpdate(false, nullptr, SuccessCallback, ErrorCallback); });
+    }
+
+    auto status_str = ParseLEAdvertisementStatus(status);
+    std::cout << "BlePublisherStatusChanged: " << status_str << std::endl;
+  }
+
   winrt::fire_and_forget BlePeripheralPlugin::AddServiceAsync(const BleService &service)
   {
     auto serviceUuid = service.uuid();
     try
     {
+
       // Build Service
       auto characteristics = service.characteristics();
       auto gattCharacteristicObjList = std::map<std::string, GattCharacteristicObject *>();
@@ -245,7 +355,7 @@ namespace ble_peripheral
         auto charProperties = characteristic.properties();
         for (flutter::EncodableValue propertyEncoded : charProperties)
         {
-          int property = static_cast<int>(std::get<int64_t>(propertyEncoded));
+          CharacteristicProperties property = std::any_cast<CharacteristicProperties>(std::get<flutter::CustomEncodableValue>(propertyEncoded));
           charParameters.CharacteristicProperties(charParameters.CharacteristicProperties() | toGattCharacteristicProperties(property));
         }
 
@@ -253,7 +363,8 @@ namespace ble_peripheral
         auto charPermissions = characteristic.permissions();
         for (flutter::EncodableValue permissionEncoded : charPermissions)
         {
-          auto blePermission = toBlePermission(static_cast<int>(std::get<int64_t>(permissionEncoded)));
+          AttributePermissions bleAttributePermission = std::any_cast<AttributePermissions>(std::get<flutter::CustomEncodableValue>(permissionEncoded));
+          auto blePermission = toBlePermission(bleAttributePermission);
           switch (blePermission)
           {
           case BlePermission::readable:
@@ -305,7 +416,8 @@ namespace ble_peripheral
           flutter::EncodableList descriptorPermissions = descriptor.permissions() == nullptr ? flutter::EncodableList() : *descriptor.permissions();
           for (flutter::EncodableValue permissionsEncoded : descriptorPermissions)
           {
-            auto blePermission = toBlePermission(static_cast<int>(std::get<int64_t>(permissionsEncoded)));
+            AttributePermissions bleAttributePermission = std::any_cast<AttributePermissions>(std::get<flutter::CustomEncodableValue>(permissionsEncoded));
+            auto blePermission = toBlePermission(bleAttributePermission);
             switch (blePermission)
             {
             case BlePermission::readable:
@@ -346,6 +458,18 @@ namespace ble_peripheral
       gattServiceProviderObject->advertisement_status_changed_token = serviceProvider.AdvertisementStatusChanged({this, &BlePeripheralPlugin::ServiceProvider_AdvertisementStatusChanged});
       serviceProviderMap.insert_or_assign(guid_to_uuid(serviceProvider.Service().Uuid()), gattServiceProviderObject);
 
+      if (serviceProvider.AdvertisementStatus() == GattServiceProviderAdvertisementStatus::Started)
+      {
+        std::cout << "Service is already advertising" << std::endl;
+      }
+      else
+      {
+        auto advertisementParameter = GattServiceProviderAdvertisingParameters();
+        advertisementParameter.IsDiscoverable(true);
+        advertisementParameter.IsConnectable(true);
+        serviceProvider.StartAdvertising(advertisementParameter);
+      }
+
       uiThreadHandler_.Post([serviceUuid]
                             { bleCallback->OnServiceAdded(serviceUuid, nullptr, SuccessCallback, ErrorCallback); });
     }
@@ -379,28 +503,28 @@ namespace ble_peripheral
   /// Advertisements Listener
   void BlePeripheralPlugin::ServiceProvider_AdvertisementStatusChanged(GattServiceProvider const &sender, GattServiceProviderAdvertisementStatusChangedEventArgs const &args)
   {
-    std::lock_guard<std::mutex> lock(cout_mutex);
     auto serviceUuid = guid_to_uuid(sender.Service().Uuid());
     if (args.Error() != BluetoothError::Success)
     {
       std::string errorStr = ParseBluetoothError(args.Error());
-      std::cout << "AdvertisementStatusChanged " << serviceUuid << ", Error " << errorStr << std::endl;
-
-      uiThreadHandler_.Post([errorStr]
-                            { bleCallback->OnAdvertisingStatusUpdate(false, &errorStr, SuccessCallback, ErrorCallback); });
-      return;
+      std::cout << "ServiceAdvertisementStatusChanged " << serviceUuid << ", Error " << errorStr << std::endl;
     }
-
-    auto argStatus = args.Status();
-    auto statusStr = AdvertisementStatusToString(argStatus);
-    std::cout << "AdvertisingStatus of service " << serviceUuid << ", changed to " << statusStr << " " << std::endl;
-
-    // Check if all services started
-    if (AreAllServicesStarted())
+    else
     {
-      uiThreadHandler_.Post([]
-                            { bleCallback->OnAdvertisingStatusUpdate(true, nullptr, SuccessCallback, ErrorCallback); });
+      std::cout << "AdvertisingStatus of service " << serviceUuid << ", changed to " << AdvertisementStatusToString(args.Status()) << " " << std::endl;
     }
+  }
+
+  void BlePeripheralPlugin::onSubscriptionUpdate(std::string deviceName, std::string deviceId, std::string characteristicId, bool subscribed)
+  {
+    /// Notify Flutter
+    uiThreadHandler_.Post([deviceName, deviceId, characteristicId, subscribed]
+                          {
+                            bleCallback->OnCharacteristicSubscriptionChange(
+                                deviceId, characteristicId, subscribed, &deviceName,
+                                SuccessCallback, ErrorCallback);
+                            // Notify subscription change
+                          });
   }
 
   /// Characteristic Listeners
@@ -442,14 +566,7 @@ namespace ble_peripheral
         {
           auto deviceInfo = co_await DeviceInformation::CreateFromIdAsync(oldClient.Session().DeviceId().Id());
           auto deviceName = winrt::to_string(deviceInfo.Name());
-          uiThreadHandler_.Post([deviceName, deviceIdArg, characteristicId]
-                                {
-                                  bleCallback->OnCharacteristicSubscriptionChange(
-                                      deviceIdArg, characteristicId, false, &deviceName,
-                                      SuccessCallback, ErrorCallback);
-                                  // Notify subscription change
-                                });
-          // Point to the local variable
+          onSubscriptionUpdate(deviceName, deviceIdArg, characteristicId, false);
         }
         catch (...)
         {
@@ -480,14 +597,7 @@ namespace ble_peripheral
         {
           auto deviceInfo = co_await DeviceInformation::CreateFromIdAsync(currentClient.Session().DeviceId().Id());
           auto deviceName = winrt::to_string(deviceInfo.Name());
-          uiThreadHandler_.Post([deviceName, deviceIdArg, characteristicId]
-                                {
-                                  bleCallback->OnCharacteristicSubscriptionChange(
-                                      deviceIdArg, characteristicId, true, &deviceName,
-                                      SuccessCallback, ErrorCallback);
-                                  // Notify subscription change
-                                });
-          // Point to the local variable
+          onSubscriptionUpdate(deviceName, deviceIdArg, characteristicId, true);
         }
         catch (...)
         {
@@ -519,7 +629,7 @@ namespace ble_peripheral
       auto bytevc = to_bytevc(charValue);
       value_arg = &bytevc;
     }
-    
+
     auto deferral = args.GetDeferral();
     auto request = co_await args.GetRequestAsync();
     if (request == nullptr)
@@ -532,7 +642,7 @@ namespace ble_peripheral
 
     std::string deviceId = ParseBluetoothClientId(args.Session().DeviceId().Id());
     int64_t offset = request.Offset();
-  
+
     uiThreadHandler_.Post([deviceId, characteristicId, offset, value_arg, deferral, request]
                           {
                             bleCallback->OnReadRequest(
@@ -630,10 +740,11 @@ namespace ble_peripheral
       gattServiceObject->obj.AdvertisementStatusChanged(gattServiceObject->advertisement_status_changed_token);
 
       // Stop advertising if started
-
       try
       {
-        if (gattServiceObject->obj.AdvertisementStatus() == GattServiceProviderAdvertisementStatus::Started)
+        auto status = gattServiceObject->obj.AdvertisementStatus();
+        // Created, Stopped, Started, Aborted, StartedWithoutAllAdvertisementData
+        if (status == GattServiceProviderAdvertisementStatus::Started)
         {
           gattServiceObject->obj.StopAdvertising();
         }
@@ -669,49 +780,51 @@ namespace ble_peripheral
       return;
     }
     oldRadioState = radioState;
-    bleCallback->OnBleStateChange(radioState == RadioState::On, SuccessCallback, ErrorCallback);
+    bool isOn = radioState == RadioState::On;
+    uiThreadHandler_.Post([isOn]
+                          { bleCallback->OnBleStateChange(isOn, SuccessCallback, ErrorCallback); });
   }
 
-  GattCharacteristicProperties BlePeripheralPlugin::toGattCharacteristicProperties(int property)
+  GattCharacteristicProperties BlePeripheralPlugin::toGattCharacteristicProperties(CharacteristicProperties property)
   {
     switch (property)
     {
-    case 0:
+    case CharacteristicProperties::kBroadcast:
       return GattCharacteristicProperties::Broadcast;
-    case 1:
+    case CharacteristicProperties::kRead:
       return GattCharacteristicProperties::Read;
-    case 2:
+    case CharacteristicProperties::kWriteWithoutResponse:
       return GattCharacteristicProperties::WriteWithoutResponse;
-    case 3:
+    case CharacteristicProperties::kWrite:
       return GattCharacteristicProperties::Write;
-    case 4:
+    case CharacteristicProperties::kNotify:
       return GattCharacteristicProperties::Notify;
-    case 5:
+    case CharacteristicProperties::kIndicate:
       return GattCharacteristicProperties::Indicate;
-    case 6:
+    case CharacteristicProperties::kAuthenticatedSignedWrites:
       return GattCharacteristicProperties::AuthenticatedSignedWrites;
-    case 7:
+    case CharacteristicProperties::kExtendedProperties:
       return GattCharacteristicProperties::ExtendedProperties;
-    case 8:
+    case CharacteristicProperties::kNotifyEncryptionRequired:
       return GattCharacteristicProperties::Notify;
-    case 9:
+    case CharacteristicProperties::kIndicateEncryptionRequired:
       return GattCharacteristicProperties::Indicate;
     default:
       return GattCharacteristicProperties::None;
     }
   }
 
-  BlePermission BlePeripheralPlugin::toBlePermission(int permission)
+  BlePermission BlePeripheralPlugin::toBlePermission(AttributePermissions permission)
   {
     switch (permission)
     {
-    case 0:
+    case AttributePermissions::kReadable:
       return BlePermission::readable;
-    case 1:
+    case AttributePermissions::kWriteable:
       return BlePermission::writeable;
-    case 2:
+    case AttributePermissions::kReadEncryptionRequired:
       return BlePermission::readEncryptionRequired;
-    case 3:
+    case AttributePermissions::kWriteEncryptionRequired:
       return BlePermission::writeEncryptionRequired;
     default:
       return BlePermission::none;
@@ -746,6 +859,27 @@ namespace ble_peripheral
       return deviceIdString.substr(pos + 1);
     }
     return deviceIdString;
+  }
+
+  std::string BlePeripheralPlugin::ParseLEAdvertisementStatus(BluetoothLEAdvertisementPublisherStatus status)
+  {
+    switch (status)
+    {
+    case BluetoothLEAdvertisementPublisherStatus::Created:
+      return "Created";
+    case BluetoothLEAdvertisementPublisherStatus::Waiting:
+      return "Waiting";
+    case BluetoothLEAdvertisementPublisherStatus::Started:
+      return "Started";
+    case BluetoothLEAdvertisementPublisherStatus::Stopped:
+      return "Stopped";
+    case BluetoothLEAdvertisementPublisherStatus::Stopping:
+      return "Stopping";
+    case BluetoothLEAdvertisementPublisherStatus::Aborted:
+      return "Aborted";
+    default:
+      return "Unknown";
+    }
   }
 
   std::string BlePeripheralPlugin::ParseBluetoothError(BluetoothError error)
@@ -790,18 +924,6 @@ namespace ble_peripheral
       }
     }
     return nullptr;
-  }
-
-  bool BlePeripheralPlugin::AreAllServicesStarted()
-  {
-    for (auto const &[key, gattServiceObject] : serviceProviderMap)
-    {
-      if (gattServiceObject->obj.AdvertisementStatus() != GattServiceProviderAdvertisementStatus::Started)
-      {
-        return false;
-      }
-    }
-    return true;
   }
 
 } // namespace ble_peripheral