+++ windows/ble_peripheral_plugin.h
@@ -83,26 +83,31 @@ namespace ble_peripheral
 
         // BluetoothLe
         Radio bluetoothRadio{nullptr};
+        BluetoothAdapter adapter{nullptr};
+        BluetoothLEAdvertisementPublisher publisher{nullptr};
+
+        void Publisher_StatusChanged(BluetoothLEAdvertisementPublisher const &sender, IInspectable const &args);
 
         winrt::fire_and_forget InitializeAdapter();
         winrt::fire_and_forget AddServiceAsync(const BleService &service);
-        GattCharacteristicProperties toGattCharacteristicProperties(int property);
-        BlePermission toBlePermission(int permission);
+        GattCharacteristicProperties toGattCharacteristicProperties(CharacteristicProperties property);
+        BlePermission toBlePermission(AttributePermissions permission);
         std::string AdvertisementStatusToString(GattServiceProviderAdvertisementStatus status);
         void disposeGattServiceObject(GattServiceProviderObject *gattServiceObject);
         void Radio_StateChanged(Radio radio, IInspectable args);
         RadioState oldRadioState = RadioState::Unknown;
         winrt::event_revoker<IRadio> radioStateChangedRevoker;
         std::string ParseBluetoothClientId(hstring clientId);
+        std::string ParseLEAdvertisementStatus(BluetoothLEAdvertisementPublisherStatus status);
 
         GattCharacteristicObject *FindGattCharacteristicObject(std::string characteristicId);
 
         void ServiceProvider_AdvertisementStatusChanged(GattServiceProvider const &sender, GattServiceProviderAdvertisementStatusChangedEventArgs const &);
         winrt::fire_and_forget SubscribedClientsChanged(GattLocalCharacteristic const &sender, IInspectable const &);
+        void onSubscriptionUpdate(std::string deviceName, std::string deviceId, std::string characteristicId, bool subscribed);
         winrt::fire_and_forget ReadRequestedAsync(GattLocalCharacteristic const &, GattReadRequestedEventArgs args);
         winrt::fire_and_forget WriteRequestedAsync(GattLocalCharacteristic const &, GattWriteRequestedEventArgs args);
         std::string ParseBluetoothError(BluetoothError error);
-        bool AreAllServicesStarted();
 
         // BlePeripheralChannel
         std::optional<FlutterError> Initialize();
@@ -114,6 +119,7 @@ namespace ble_peripheral
         std::optional<FlutterError> RemoveService(const std::string &service_id);
         std::optional<FlutterError> ClearServices();
         ErrorOr<flutter::EncodableList> GetServices();
+        ErrorOr<flutter::EncodableList> GetSubscribedClients();
         std::optional<FlutterError> StartAdvertising(
             const flutter::EncodableList &services,
             const std::string *local_name,